suppression des parties inutiles, implementation de firebase à la page login en cours
