+++ src/components/form/index.tsx
@@ -1,70 +1,67 @@
 import {
-  chakra,
-  Flex,
-  Grid,
-  useDisclosure,
-  useUpdateEffect,
-  HTMLChakraProps,
-  FormControl,
-  Input,
-  FormLabel,
-  Spacer,
-  Link,
-  Box,
+    chakra,
+    Flex,
+    Grid,
+    useDisclosure,
+    useUpdateEffect,
+    HTMLChakraProps,
+    FormControl,
+    Input,
+    FormLabel,
+    Spacer,
+    Link,
+    Box, Button,
 
 } from "@chakra-ui/react"
 import React from "react"
-import { useForm } from "react-hook-form";
-import { useViewportScroll } from "framer-motion";
+import {useForm} from "react-hook-form";
+import {useViewportScroll} from "framer-motion";
+import firebase from "../../utils/client";
+import * as url from "url";
+interface LoginProps {
+    email: string
+}
 
-  
-
-    
 
 const Form_exp = (props: HTMLChakraProps<"form">) => {
-    const { register, formState: { errors }, handleSubmit } = useForm();
-  const onSubmit = (data) => alert(JSON.stringify(data));
-  const mobileNav = useDisclosure();
-  const mobileNavBtnRef = React.useRef<HTMLButtonElement>();
+    var actionCodeSettings = {
+        url: "http://localhost:3000/login",
+        handleCodeInApp: true
+    };
+    const login = async (props: LoginProps) => {
+        try {
+            const user = await firebase.auth().sendSignInLinkToEmail(props.email, actionCodeSettings);
+            if (user != undefined) {
+                window.localStorage.setItem('emailForSignIn', props.email);
+            }else{
+                window.localStorage.setItem('nop', props.email);
+            }
+        } catch (e) {
+            console.log(e);
+        }
+    }
 
-  useUpdateEffect(() => {
-      mobileNavBtnRef.current?.focus();
-  }, [mobileNav.isOpen]);
-  const ref = React.useRef<HTMLFormElement>()
-  const [y, setY] = React.useState(0)
-  const { height = 0 } = ref.current?.getBoundingClientRect() ?? {}
 
-  const { scrollY } = useViewportScroll()
-  React.useEffect(() => {
-      return scrollY.onChange(() => setY(scrollY.get()))
-  }, [scrollY])
+    const {register, formState: {errors}, handleSubmit} = useForm();
+    const onSubmit = async (data) => {
+        await login({email: data.email})
+    };
 
-  return (
-    <Grid p="0.5em" w="25em" border='1px' flex-direction='column' mt="2em" >
-    <form onSubmit={handleSubmit(onSubmit)}>
-    <Flex  >
-    <FormControl id="email">
-        <FormLabel>Email adress 👤</FormLabel>
-        <Input {...register("email")}/>
-    </FormControl>
-    </Flex>
-    <br/>
-    <Flex >
-    <FormControl id="password">
-        <Box>
-        <FormLabel>Password 🔑 <Link color="blue" float="right">Forgot Password</Link></FormLabel>
-        </Box>
-        <Input {...register("password")} />
-        </FormControl>
-        
-    </Flex>
-    <br/>
-    <Flex text-align="center">
-        <Input type="submit" color="white" value ="Sign In" bgColor="black"  />
-    </Flex>
-    </form>
-</Grid> 
-  );
+    return (
+        <Grid p="0.5em" w="25em" border='1px' flex-direction='column' mt="2em">
+            <form onSubmit={handleSubmit(onSubmit)}>
+                <Flex>
+                    <FormControl id="email">
+                        <FormLabel>Email adress 👤</FormLabel>
+                        <Input  {...register("email")}/>
+                    </FormControl>
+                </Flex>
+                <Flex text-align="center">
+                    <Input type="submit" color="white" value="Sign In" bgColor="black"/>
+                </Flex>
+            </form>
+        </Grid>
+    );
 }
 
 export default Form_exp
